requirements.txt
asgiref==3.7.2
blinker==1.7.0
certifi==2023.11.17
click==8.1.7
dill==0.3.8
Django==5.0.2
django-cors-headers==4.3.1
dnspython==2.5.0
Flask==3.0.2
Flask-Cors==4.0.0
groupdocs-comparison-cloud==23.12
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.3.2
MarkupSafe==2.1.5
nltk==3.8.1
numpy==1.26.4
passlib==1.7.4
pymongo==4.6.1
python-dateutil==2.8.2
python-dotenv==1.0.1
regex==2023.12.25
route==0.1.4
scikit-learn==1.4.0
scipy==1.12.0
Send2Trash==1.8.2
setuptools==69.0.3
six==1.16.0
sqlparse==0.4.4
threadpoolctl==3.2.0
tqdm==4.66.1
urllib3==2.2.1
Werkzeug==3.0.1
wheel==0.42.0
